requirements.txt
Flask==3.1.3
mysql-connector-python==9.6.0
reportlab==4.4.10
python-dotenv==1.2.1
